refactor(data_processor): 优化数据处理和类型转换
- 在数据处理中添加未来行为的显式选择，以避免潜在的类型转换问题
- 使用 infer_objects(copy=False) 优化类型转换，减少不必要的对象复制
- 确保在测试中正确处理 NaN 值，提高测试的准确性和可靠性

+++ tests/test_nwc_calculator.py
@@ -68,7 +68,9 @@ def test_calculate_delta_nwc_success(calculator):
     # Yr2: 535 - 470 = 65
     # Yr3: 600 - 535 = 65
     expected_delta = pd.Series([Decimal('NaN'), Decimal('65'), Decimal('65')])
-    pd.testing.assert_series_equal(delta_nwc, expected_delta, check_index=False, check_dtype=False)
+    # Convert expected_delta to float to ensure NaN is np.nan for comparison
+    # delta_nwc from .diff() is already float with np.nan
+    pd.testing.assert_series_equal(delta_nwc, expected_delta.astype(float), check_index=False, check_dtype=False)
 
 def test_calculate_delta_nwc_with_nan(calculator):
     nwc = pd.Series([Decimal('470'), Decimal('NaN'), Decimal('600')])
